Optimize role code
